generator/requirements.txt
httpx
requests
python-dotenv